use singleton for config
