+++ packages/x-charts/src/internals/index.ts
@@ -1,0 +1,1 @@
+export * from './configInit';
